API UPDATE & SERVICE CONNECTION EXAMPLES
- Api update
- Swagger compatibility
- Get and Post on other services example
- DTOs fix

+++ common/src/main/kotlin/it/polito/master/ap/group6/ecommerce/common/dtos/TransactionDTO.kt
@@ -4,7 +4,7 @@ import java.util.*
 
 data class TransactionDTO(
     val user: UserDTO? = null,
-    val amount: Int? = null,
+    val amount: Float? = null,
     val timestamp: Date? = null,
     val causal: String? = null,
     val purchases: List<PurchaseDTO>? = null
